requirements/staging.txt
-r base.txt
django-debug-toolbar==1.6

